requirement.txt
ultralytics==8.0.203
transformers==4.37.2
torch==2.1.2
torchvision==0.16.1
supervision==0.16.0
scikit-learn==1.3.2
umap-learn==0.5.5
tqdm==4.66.1
Pillow==10.2.0
more-itertools==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
